requirements.txt
anyio==4.3.0
APScheduler==3.10.4
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.5
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jaraco.classes==3.3.1
jeepney==0.8.0
keyring==24.3.0
more-itertools==10.2.0
psutil==5.9.8
pycparser==2.21
python-telegram-bot==21.0.1
pytz==2023.4
PyYAML==6.0.1
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.25
typing_extensions==4.10.0
tzlocal==5.2
